Refactorizacion de Home
